+++ Examples/SwiftUI/Superwall-SwiftUI/Services/StoreKitService.swift
@@ -5,7 +5,7 @@
 //  Created by Yusuf Tör on 05/04/2022.
 //
 
-// Uncomment if you're implementing the SubscriptionController in SuperwallService.swift:
+// Uncomment if you're implementing the PurchaseController in SuperwallService.swift:
 /*
 import StoreKit
 import SuperwallKit