+++ Examples/UIKit-Swift/Superwall-UIKit-Swift/Services/StoreKitService.swift
@@ -5,8 +5,8 @@
 //  Created by Yusuf Tör on 05/04/2022.
 //
 
-// Uncomment if you're implementing the SubscriptionController in SuperwallService.swift:
-/*
+// Uncomment if you're implementing the PurchaseController in SuperwallService.swift:
+
 import StoreKit
 import SuperwallKit
 
@@ -129,4 +129,3 @@ extension StoreKitService: SKPaymentTransactionObserver {
     }
   }
 }
-*/